1 修复休眠开关按键死机问题；2 修复待机1.7uA,正常广播100uA,连接上ble工作470uA

+++ 5.0.4/sdk/platform/driver/battery/battery.h
@@ -57,6 +57,6 @@
  * @return Battery level. 0 - 100%
  ****************************************************************************************
  */
-uint8_t battery_get_lvl(uint8_t batt_type);
+uint16_t battery_get_lvl(uint8_t batt_type);
 
 #endif // _BATTERY_H_
